+++ host/usb/include/usb/usb_host.h
@@ -4,10 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/*
-Warning: The USB Host Library API is still a beta version and may be subject to change
-*/
-
 #pragma once
 
 #include <stdint.h>
@@ -24,6 +20,8 @@ extern "C" {
 #endif
 
 // ------------------------------------------------- Macros and Types --------------------------------------------------
+
+#define REMOTE_WAKE_HAL_SUPPORTED (1)  // Define for class drivers that this version of usb component supports the remote wakeup HAL API.
 
 // ----------------------- Handles -------------------------