+++ host/usb/src/hcd_dwc.c
@@ -893,7 +893,6 @@ static hcd_port_event_t _intr_hdlr_hprt(port_t *port, usb_dwc_hal_port_event_t h
         port->flags.conn_dev_ena = 0;
         break;
     }
-#ifdef REMOTE_WAKE_HAL_SUPPORTED
     case USB_DWC_HAL_PORT_EVENT_REMOTE_WAKEUP: {
         ESP_EARLY_LOGD(HCD_DWC_TAG, "Remote wakeup generated from device");
         // Port must have been previously suspended to start processing remote wakeup signaling
@@ -902,7 +901,6 @@ static hcd_port_event_t _intr_hdlr_hprt(port_t *port, usb_dwc_hal_port_event_t h
         }
         break;
     }
-#endif // REMOTE_WAKE_HAL_SUPPORTED
     default: {
         abort();
         break;
@@ -1200,9 +1198,6 @@ static inline bool _is_fifo_config_by_bias(const usb_dwc_hal_fifo_config_t *cfg)
 /**
  * @brief Gate internal clock
  *
- * @note HAL functions used in this function were backported together with remote wakeup changes, we are guarding the
- *       presence of both using the same public define
- *
  * @param[in] port Pointer to the port object
  * @param[in] enable enable/disable internal clock gating
  * @return True internal clk successfully gated/un-gated
@@ -1210,7 +1205,6 @@ static inline bool _is_fifo_config_by_bias(const usb_dwc_hal_fifo_config_t *cfg)
  */
 static inline bool _internal_clk_gate(port_t *port, bool enable)
 {
-#ifdef REMOTE_WAKE_HAL_SUPPORTED
     // Stop PHY Clock and gate HCLK
     usb_dwc_hal_pwr_clk_internal_clock_gate(port->hal, enable);
 
@@ -1227,9 +1221,6 @@ static inline bool _internal_clk_gate(port_t *port, bool enable)
         return true;
     }
     return false;
-#else
-    return true;
-#endif // REMOTE_WAKE_HAL_SUPPORTED
 }
 
 // ---------------------- Commands -------------------------