+++ host/usb/test/target_test/usb_host/main/ctrl_client.h
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2015-2025 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2015-2026 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -13,6 +13,4 @@ typedef struct {
 
 void ctrl_client_async_seq_task(void *arg);
 
-#ifdef REMOTE_WAKE_HAL_SUPPORTED
 void ctrl_client_remote_wake_task(void *arg);
-#endif // REMOTE_WAKE_HAL_SUPPORTED